requirements.txt
numpy==1.26.4
pygame==2.5.2
pygame-widgets==1.1.5
PyOpenGL==3.1.7
